+++ samples/file-encryptor/enclave/common/file-encryptor.conf
