+++ samples/file-encryptor/enclave/common/trace.h
