+++ samples/file-encryptor/enclave/mbedtls/keys.cpp
@@ -12,7 +12,7 @@
 #include <mbedtls/rsa.h>
 #include <mbedtls/sha256.h>
 
-#include "common.h"
+#include "../common/trace.h"
 #include "encryptor.h"
 
 #define ENCRYPT_OPERATION true
